% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{calculate_growth_scenarios}
\alias{calculate_growth_scenarios}
\title{Calculate Population Growth Scenarios}
\usage{
calculate_growth_scenarios(pop_data)
}
\arguments{
\item{pop_data}{Input population data}
}
\description{
Calculate Population Growth Scenarios
}
