% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic.R
\name{county_population_projection_plot}
\alias{county_population_projection_plot}
\title{Create County Population Projection Plot}
\usage{
county_population_projection_plot(df)
}
\arguments{
\item{df}{Data frame containing projection data}
}
\description{
Create County Population Projection Plot
}
