% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_editions.R
\name{plot_editions_simple}
\alias{plot_editions_simple}
\title{Simple plot to compare the different editions.}
\usage{
plot_editions_simple(pop)
}
\description{
Simple plot to compare the different editions.
}
