% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{pop19}
\alias{pop19}
\title{pop19}
\format{
A data frame with columns including FIPS, County, Year, Type(Total), and Population.
\describe{#'
\item{\code{FIPS}}{FIPS code of each county}
\item{\code{County}}{Name  of county}
\item{\code{Year}}{Year of Projection (2000-2035)}
\item{\code{Type}}{Estimate or Projection}
\item{\code{Population}}{Population number}
}
}
\usage{
pop19
}
\description{
Estimate and projection totals
}
\examples{
data(pop19)
head(pop19)
}
\keyword{datasets}
