% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop19_gender}
\alias{pop19_gender}
\title{Population Data by Gender from 2019 Edition (pop19_gender)
This dataset came from the SC Office of Revenue and Fiscal Affairs. It contains data for the Total, Male, and Female populations for each county. Population estimates from year 2000 to 2018 and population projections from 2019 to 2035 are included.}
\format{
A data frame with columns:
\describe{
\item{\code{FIPS}}{FIPS code of each county}
\item{\code{County}}{Name  of county}
\item{\code{Year}}{Year of Projection/Estimate (2000-2035)}
\item{\code{Type}}{Estimate or Projection}
\item{\code{Gender}}{Male , Female or Total}
\item{\code{Population}}{Population number}
}
}
\usage{
pop19_gender
}
\description{
Population Data by Gender from 2019 Edition (pop19_gender)
This dataset came from the SC Office of Revenue and Fiscal Affairs. It contains data for the Total, Male, and Female populations for each county. Population estimates from year 2000 to 2018 and population projections from 2019 to 2035 are included.
}
\examples{
data(pop19_gender)
head(pop19_gender)
}
\keyword{datasets}
