% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop22}
\alias{pop22}
\title{Population totals from the 2022 edition (pop22)
This dataset came from the SC Office of Revenue and Fiscal Affairs.
Estimate and projection totals}
\format{
A data frame with columns including FIPS, County, Year, Type(Total), and Population.
\describe{#'
\item{\code{FIPS}}{FIPS code of each county}
\item{\code{County}}{Name  of county}
\item{\code{Year}}{Year of Projection (2000-2038)}
\item{\code{Type}}{Estimate or Projection}
\item{\code{Population}}{Population number}
}
}
\usage{
pop22
}
\description{
Population totals from the 2022 edition (pop22)
This dataset came from the SC Office of Revenue and Fiscal Affairs.
Estimate and projection totals
}
\examples{
data(pop22)
head(pop22)
}
\keyword{datasets}
