% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop22_cohort}
\alias{pop22_cohort}
\title{Population by cohort from the 2022 edition (pop22_cohort)
This dataset came from the SC Office of Revenue and Fiscal Affairs.
A transformed and summarized population data, distinguishing between Estimate and Projections, reformatted for clarity, and adding a state-wide summary.}
\format{
Data frame including FIPS, County, Year, Type, Gender and Population.
\describe{
\item{\code{AGE GROUP}}{Age Group of Population,  "0 TO 4"   "5 TO 9"   "10 TO 14" "15 To 19" "20 TO 24" "25 TO 29" "30 TO 34" "35 TO 39" "40 TO 44" "45 TO 49" "50 TO 54" "55 TO 59" "60 TO 54"
"65 TO 69" "70 TO 74" "75 TO 79" "80 TO 84" "85+" }
\item{\code{FIPS}}{FIPS code of each county}
\item{\code{County}}{Name  of county}
\item{\code{Year}}{Year of Projection/Estimate (2010-2038)}
\item{\code{Type}}{Estimate or Projection}
\item{\code{Gender}}{Male , Female or Total}
\item{\code{Population}}{Population number}
}
}
\usage{
pop22_cohort
}
\description{
Population by cohort from the 2022 edition (pop22_cohort)
This dataset came from the SC Office of Revenue and Fiscal Affairs.
A transformed and summarized population data, distinguishing between Estimate and Projections, reformatted for clarity, and adding a state-wide summary.
}
\examples{
data(pop22_cohort)
head(pop22_cohort)
}
\keyword{datasets}
