% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop22_gender}
\alias{pop22_gender}
\title{Population by gender from the 2022 edition (pop22_gender)
This dataset came from the SC Office of Revenue and Fiscal Affairs.
Tidy Format of (Nov 2022 Data)}
\format{
A tidy data frame after reshaped to \emph{long form} by creating new columns for Year, Type (Estimate or Projection), Gender (Total, Male, or Female), and Population.
\describe{
\item{\code{FIPS}}{FIPS code of each county}
\item{\code{County}}{Name  of county}
\item{\code{Year}}{Year of Projection/Estimate (2010-2038)}
\item{\code{Type}}{Estimate or Projection}
\item{\code{Gender}}{Male , Female or Total}
\item{\code{Population}}{Population number}
}
}
\usage{
pop22_gender
}
\description{
Population by gender from the 2022 edition (pop22_gender)
This dataset came from the SC Office of Revenue and Fiscal Affairs.
Tidy Format of (Nov 2022 Data)
}
\examples{
data(pop22_gender )
 head(pop22_gender )
}
\keyword{datasets}
