% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_proj}
\alias{pop_proj}
\title{pop_proj Results of extending the SC RFA projections
This table includes detailed results of the methods used to extend the projections (for both editions of the SC RFA projections).}
\format{
A data frame with columns:
\describe{
\item{\code{Period}}{Original (from SC RFA) or Extended (from the methods used to extend the projections)}
\item{\code{Scenario}}{Estimate, Moderate, or High}
\item{\code{Edition}}{2019 or 2022}
\item{\code{County}}{Name of the county}
\item{\code{Year}}{Year of Projection (2000-2070)}
\item{\code{Type}}{Estimate, SC RFA Projection, or Extrapolated Projection}
\item{\code{Population}}{Population number}
\item{\code{T_0}}{Initial year for this "scenario"}
\item{\code{T_1}}{Final year for this "scenario"}
\item{\code{Pop_0}}{Initial population for this "scenario"}
\item{\code{Pop_1}}{Final population for this "scenario"}
\item{\code{Growth_linear}}{Population growth for this "scenario," annual change in population}
\item{\code{Growth_exp}}{Exponential growth rate for this "scenario"}
\item{\code{Growth_moderate}}{Annual change  in population under the moderate growth scenario (linear growth with minimum set to zero)}
\item{\code{Growth_exp_floor}}{Minimum growth rate for the High scenario (the state-wide average*1.1)}
\item{\code{Growth_hi}}{Exponential growth rate for this "scenario," floor set to at least the state average growth rate}
\item{\code{Type2}}{Relative growth in population under the moderate growth scenario (2019 edition)}
}
\usage{
pop_proj
}
\description{
pop_proj Results of extending the SC RFA projections
This table includes detailed results of the methods used to extend the projections (for both editions of the SC RFA projections).
}
\examples{
data(pop_proj)
head(pop_proj)
}
\keyword{datasets}
