% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{pop_proj_relative}
\alias{pop_proj_relative}
\title{pop_proj_relative: Relative Growth Projections for Different Counties}
\format{
The dataset is a data frame with the following columns:
\describe{
\item{\code{County}}{Name of the county}
\item{\code{Year}}{Year of Projection (2000-2038)}
\item{\code{High}}{Population in the high growth scenario}
\item{\code{Moderate}}{Population in the moderate growth scenario}
\item{\code{Driver_growth_mod}}{Relative growth in population under the moderate growth scenario}
\item{\code{Driver_growth_hi}}{Relative growth in population under the high growth scenario}
}

The formulas used in the dataset are as follows:
\itemize{
\item Moderate Growth Scenario (Moderate):
\deqn{Growth\_{moderate} = \max(Growth\_{linear}, 0)}
Here, \eqn{Growth\_{linear}} represents the average annual increase in population.

\item High Growth Scenario (High):
\deqn{Growth\_{hi} = Growth\_{exp\_floor} \times 1.10}
\eqn{Growth\_{exp}} is the exponential growth rate. \eqn{Growth\_{exp\_floor}} is set to at least the state average growth rate.

\item Driver Growth Factors:
\deqn{Driver\_{growth\_mod} = \frac{Moderate Population\ in\ Year\ X}{Moderate Population\ in\ 2021}}
\deqn{Driver\_{growth\_hi} = \frac{High Population\ in\ Year\ X}{High Population\ in\ 2021}}
}
}
\usage{
pop_proj_relative
}
\description{
This dataset provides relative growth projections for different counties under moderate and high growth scenarios. It includes calculations for growth rates and driver growth factors under each scenario.
}
\examples{
data(pop_proj_relative)
head(pop_proj_relative)
}
\keyword{datasets}
