% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyramid.R
\name{pyramid_plot}
\alias{pyramid_plot}
\title{Create Population Pyramid Plot}
\usage{
pyramid_plot(df)
}
\arguments{
\item{df}{Data frame containing population data}
}
\description{
Create Population Pyramid Plot
}
\details{
See https://rfortherestofus.com/2024/07/population-pyramid-part-1 for inspiration.
}
